small fixes

+++ server/public/scripts/client.js
@@ -78,16 +78,21 @@ function getTask() {
 
 function deleteTask() {
     console.log('in deleteTask', $(this).data('id'));
-    let taskId = $(this).data('id')
-    $.ajax({
-        method: 'DELETE',
-        url: '/tasks?id=' + taskId
-    }).then(function (response) {
-        getTask();
-    }).catch(function (err) {
-        console.log(err);
-        alert('error deleting task');
-    })
+    if (confirm('Would you like to delete?') == true) {
+        let taskId = $(this).data('id')
+        $.ajax({
+            method: 'DELETE',
+            url: '/tasks?id=' + taskId
+        }).then(function (response) {
+            getTask();
+        }).catch(function (err) {
+            console.log(err);
+            alert('error deleting task');
+        })
+    }
+    else {
+        return
+    }
 }
 
 function updateTask() {
